+++ utils/runAsync.js
@@ -1,0 +1,11 @@
+// Runs a query and returns a promise
+const runAsync = (query, params = []) => {
+  return new Promise((resolve, reject) => {
+    db.run(query, params, function (err) {
+      if (err) reject(err);
+      else resolve(this);
+    });
+  });
+};
+
+module.exports = runAsync;
